+++ Planarian/Planarian.Model/Database/Entities/RidgeWalker/EntranceHydrologyTag.cs
@@ -1,0 +1,12 @@
+using Planarian.Model.Shared.Base;
+
+namespace Planarian.Model.Database.Entities.RidgeWalker;
+
+public class EntranceHydrologyTag : EntityBase
+{
+    public string TagTypeId { get; set; }
+    public string EntranceId { get; set; }
+
+    public TagType TagType { get; set; }
+    public Entrance Entrance { get; set; }
+}
